Consolidate helper functions into single file
